+++ src/org/labkey/test/util/exp/SampleTypeAPIHelper.java
@@ -75,7 +75,7 @@ public class SampleTypeAPIHelper
      * @return A map of containing sample names and their corresponding row ids.
      * @throws Exception Because this uses the Select Rows Command it can throw a few different type of exceptions.
      */
-    public static Map<String, Long> getRowIdsForSamples(String containerPath, String sampleTypeName, List<String> sampleNames) throws IOException, CommandException
+    public static Map<String, Integer> getRowIdsForSamples(String containerPath, String sampleTypeName, List<String> sampleNames) throws IOException, CommandException
     {
 
         Connection connection = WebTestHelper.getRemoteApiConnection();
@@ -85,13 +85,13 @@ public class SampleTypeAPIHelper
 
         SelectRowsResponse response = cmd.execute(connection, containerPath);
 
-        Map<String, Long> rowIds = new HashMap<>();
+        Map<String, Integer> rowIds = new HashMap<>();
 
         for(Map<String, Object> row : response.getRows())
         {
             Object name = row.get("Name");
             Object value = row.get("RowId");
-            rowIds.put(name.toString(), Long.parseLong(value.toString()));
+            rowIds.put(name.toString(), Integer.parseInt(value.toString()));
         }
 
         // Check that the names returned from the query match the names sent in.
@@ -108,7 +108,7 @@ public class SampleTypeAPIHelper
      * @deprecated Use {@link #getRowIdsForSamples(String, String, List)}
      */
     @Deprecated(since = "22.4")
-    public static Map<String, Long> getSampleIdFromName(String folder, String sampleTypeName, List<String> sampleNames) throws IOException, CommandException
+    public static Map<String, Integer> getSampleIdFromName(String folder, String sampleTypeName, List<String> sampleNames) throws IOException, CommandException
     {
         return getRowIdsForSamples(folder, sampleTypeName, sampleNames);
     }
